Database, app, and settings changes
I've attached a blank db file with this commit and changed several files
to correct errors I was having. The next step is to get the admin site up and
running and to create a superuser for all of us to use. I'll explain that
when I get to it.

Once I have the admin site, I'll work on populating the database.