Template Acquisition
Based on the files sent to me by Andres, I have implemented the
templates and new static files in a way suitable for django. There were
some new conventions I had to learn about settings and urls in order to
do this. However, the base of the site now has a face to go with it.