Media Update
The database is complete, my image and sound upload script is done and
we now have all waterfowl entries completed for testing. Of note for
future commits, some of the files were improperly named and so the
script didn't work at first. I did not input a conditional to skip over
that this time because I want to know if there are birds that aren't
getting their proper files.

Please note I removed the actual media directory to a different
directory on my machine in order to cut down on the size of the
repository. That's why the data is missing.