Server Restart Nuisance and Template Update
Before, if you tried to load a bird page without it having sound or
image, it would crash the whole page. That's fixed now.

Unfortunately, I've discovered something that is a bit of a nuisance.
Updating the database while the server is running will not register the
changes in the template. If I try to add Test Bird 2, for example, it
won't immediately show up in the front page list like it should.
Instead, I have to restart the server for it to appear.

It's more of a nuisance than anything, but I'm going to try getting it
fixed sometime before the presentation.